REF (string): de-duplicate str_isfoo methods (#59705)
